requirements.txt
crcmod==1.7
paho-mqtt==1.5.1
